pkg/k8s/endpointsynchronizer: move package to pkg/k8s/watchers
