+++ pkg/k8s/watchers/endpointsynchronizer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package endpointsynchronizer
+package watchers
 
 import (
 	"context"
@@ -34,6 +34,11 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+)
+
+const (
+	// subsysEndpointSync is the value for logfields.LogSubsys
+	subsysEndpointSync = "endpointsynchronizer"
 )
 
 // EndpointSynchronizer currently is an empty type, which wraps around syncing
